Fix compiler warnings in playsim and fonts
p_acs.cpp: 8× -Wsign-compare (int/int32_t vs size_t)
p_effect.cpp: 1× -Wsign-compare (uint32_t vs int)
p_lnspec.cpp: 4× -Wsign-compare (int vs size_t)
p_map.cpp: 1× -Wint-in-bool-context (enum to bool)
p_mobj.cpp: 1× -Wsign-compare (unsigned int vs int)
teaminfo.cpp: 1× -Wsign-compare (int vs unsigned int)
v_text.cpp: 4× -Wchar-subscripts (char used as array index)

+++ src/playsim/p_effect.cpp
@@ -124,7 +124,7 @@ static void FreeParticle(FLevelLocals* Level, particle_t* particle)
 		assert(next->tprev == pindex);
 		next->tprev = particle->tprev;
 	}
-	if (Level->OldestParticle == pindex)
+	if (Level->OldestParticle == (uint32_t)pindex)
 	{
 		assert(tnext == NO_PARTICLE);
 		Level->OldestParticle = particle->tprev;
